Encapsulate bus estimation elements
State for each of the dropdown elements has been lifted to this parent
busEstimate component in order to keep all the components synced when an
option is selected. Each dropdown is a reusable component that takes a
display string, an array of contents that will populate the dropdown,
and a callback to trigger state change in the parent component.